Move tree renderer into separate file
